Fix Cannot convert [array()] to EagerTensor of dtype int64
While running the model.prepare_tf_dataset() method,
it raises the error below:
```
TypeError: Cannot convert [array([322.,   1.])] to EagerTensor of dtype int64
```

This happens, in  "DataCollatorForSeq2Seq" function when we are try
to convert the labels to tensors. While converting the labels to tensors,
the labels can be in the format of list of list or list of ndarrays.
There is no problem converting the list of list lables. When the values are
list of ndarray(like below) we need to handle it separately.

```
[array([322.,   1.])]
```

so the exception raises while trying to convert this label to tensors using
below code.

```
batch["labels"] = tf.constant(batch["labels"], dtype=tf.int64)
```

This would have worked fine if the value inside the array is int,
the problem is with the dtype change here.

So we need to convert the list of ndarray to ndarray before passing it
to the tf.constant() funciton.

This commit makes sure labels are list or numpy array and
converts them accordingly.